Extracted dispatchers to injectable class thus fixing viewModel tests
