+++ app/src/test/java/io/gnosis/safe/TestAppDispatchers.kt
@@ -1,0 +1,6 @@
+package io.gnosis.safe
+
+import io.gnosis.safe.ui.base.AppDispatchers
+import kotlinx.coroutines.test.TestCoroutineDispatcher
+
+val appDispatchers = AppDispatchers(TestCoroutineDispatcher(), TestCoroutineDispatcher())
